+++ test/plugin/add_metadata_test.rb
@@ -50,7 +50,7 @@ describe Shrine::Plugins::AddMetadata do
       end
     end
 
-    describe "with skip_nil option" do
+    describe "with :skip_nil option" do
       it "does not ignore not nil values" do
         @shrine.add_metadata(:custom, skip_nil: true) { false }
         metadata = @uploader.extract_metadata(fakeio)
@@ -60,7 +60,7 @@ describe Shrine::Plugins::AddMetadata do
       it "ignores nil" do
         @shrine.add_metadata(:custom, skip_nil: true) { nil }
         metadata = @uploader.extract_metadata(fakeio)
-        assert !metadata.has_key?("custom")
+        refute metadata.has_key?("custom")
       end
     end